% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grids.R
\name{check_grids}
\alias{check_grids}
\title{Summarize and check evaluation grids for the learnitgrid Shiny app}
\usage{
check_grids(dir, save.log = TRUE, save.rds = TRUE)
}
\arguments{
\item{dir}{The path to the evaluation grids}

\item{save.log}{Should a log file be saved (yes by default)?}

\item{save.rds}{Should an RDS file with the data be saved (yes by default)?}
}
\value{
A data frame with the results of the evaluation grids check for the
learnitgrid Shiny app.
}
\description{
Check all evaluation grids in a correction set and computes
the required statistics for the summary page of the learnitgrid shiny app.
}
