% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_cache.R
\name{copy_cache}
\alias{copy_cache}
\alias{CC}
\title{Copy data cache from a reference folder into an RStudio project}
\usage{
copy_cache()

CC(only_if_copied = FALSE, cache_data = cache_data)
}
\arguments{
\item{only_if_copied}{If \code{TRUE}, the cache is only deleted if it was copied}

\item{cache_data}{The list of copied files, returned by \code{copy_cache()}}
}
\value{
The list of copied files is returned invisibly.
}
\description{
When GitHub repositories for a Classroom assessment are cloned, the data
cache is not present. For large datasets, it may take a long time to get
the data from an URL for each repository. By using a single version of the
data cache, on can populated the corresponding folder in each repository.
The \code{CC()} function can be used to remove the data from the cache.
}
