% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_rmd}
\alias{correct_rmd}
\title{Make sure that Rmd/Qmd chunk labels are written without spaces}
\usage{
correct_rmd(rmd)
}
\arguments{
\item{rmd}{Character string with the content of a Rmd/Qmd file.}
}
\value{
The same Rmd/Qmd content, but with "educated" chunk labels.
}
\description{
Make sure that Rmd/Qmd chunk labels are written without spaces
}
