% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_context.R
\name{create_context}
\alias{create_context}
\title{Create a context object for a correction set}
\usage{
create_context(
  correction,
  base_corr_dir,
  base_templ_dir,
  base_repos_dir,
  repositories,
  assignments,
  github_url,
  branch
)
}
\arguments{
\item{correction}{The correction set}

\item{base_corr_dir}{The base directory for correction sets}

\item{base_templ_dir}{The base directory for correction templates}

\item{base_repos_dir}{The vase directory for GitHub repositories to correct}

\item{repositories}{The repositories}

\item{assignments}{The assignments}

\item{github_url}{The GitHub URL (base part)}

\item{branch}{The GitHub branch concerned by this correction}
}
\value{
A list with the context information for the learnitgrid Shiny app.
}
\description{
Create a list that contains context for a given correction set (must be
recalculated when a different correction set/project is selected).
}
