% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_object.R
\name{df_structure}
\alias{df_structure}
\alias{digest}
\alias{object_attr}
\alias{object_part}
\alias{object_str}
\alias{chart_structure}
\title{Transformation functions for creating reference or result objects for testing
progress of the students in their project.}
\usage{
df_structure(object, ...)

digest(object, algo = "md5", ...)

object_attr(object, attrib = "class", ...)

object_part(object, part = "x", ...)

object_str(object, part = "x", ...)

chart_structure(object, arg = "", ...)
}
\arguments{
\item{object}{The object to transform}

\item{...}{Further arguments (not used for now)}

\item{algo}{algorithm to use for digest, \code{"md5"} by default}

\item{attrib}{The attribute(s) to record from the object}

\item{part}{The part(s) to record from the object (list items)}

\item{arg}{Additional argument. Currently, could be \code{""} (by default) to
record everything, or \code{"no.data"} to ignore data in ggplot objects.}
}
\value{
\code{df_structure()} returns names, attributes "label" and "units",
number of row and columns, classes, if there are missing data and comment in
a data frame. \code{digest()} returns a hash of the object. \code{object_attr()}
returns the attribute(s) of the object. \code{object_part()} returns the part(s)
of an object (list items). \code{object_str()} returns a str() representation of
the object parts.
}
\description{
Transformation functions for creating reference or result objects for testing
progress of the students in their project.
}
\examples{
# not yet
}
