% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_files.R
\name{diff_files}
\alias{diff_files}
\title{Launch a Shiny application to compare original and solution files}
\usage{
diff_files()
}
\value{
Launches a Shiny application.
}
\description{
This function starts a Shiny app that allows users to select an "_original" file
and automatically compares it with its corresponding "_solution" file using the \code{diffr} package.
}
\examples{
#diff_files()
}
