% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_path_create}
\alias{dir_path_create}
\alias{dir_path_check}
\alias{file_path_check}
\title{Make sure a directory exists, or create it}
\usage{
dir_path_create(...)

dir_path_check(...)

file_path_check(...)
}
\arguments{
\item{...}{The successive folders that make the path.}
}
\value{
A \strong{path} object is returned invisibly.
}
\description{
Make sure a directory exists, or create it
}
\section{Functions}{
\itemize{
\item \code{dir_path_check()}: Check that a directory exists.

\item \code{file_path_check()}: Check that a file exists.

}}
\examples{
test_path <- dir_path_create(tempdir(), "dir_path_create_test", "subfolder")
test_path
dir.exists(test_path) # Should be TRUE
dir_path_check(test_path) # Return the path only if it exists
# Remove
unlink(test_path)
unlink(dirname(test_path))
}
