% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt_solutions}
\alias{encrypt_solutions}
\alias{decrypt_solutions}
\alias{set_key}
\title{Encrypt solution files into \code{..._solution.xxx.aes} files}
\usage{
encrypt_solutions(key = NULL, error = TRUE)

decrypt_solutions(key = NULL, error = TRUE)

set_key()
}
\arguments{
\item{key}{The key to use for encryption/decryption. If not provided, it is
asked}

\item{error}{If \code{TRUE}, an error is generate in case the file encryption or
decryption fails.}
}
\value{
\code{NULL} invisibly.
}
\description{
These functions manage encryption and decryption of solution files. The
password must be defined in an option \code{learnitgrid.key} or an environment
variable \code{LEARNITGRID_KEY} under a digest form.
}
