% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_git_stats.R
\name{get_git_stats}
\alias{get_git_stats}
\title{Summarize the git stats for one or more git repositories}
\usage{
get_git_stats(
  dir,
  exclude_authors = "github-classroom[bot]",
  type = "all",
  tz = "UTC"
)
}
\arguments{
\item{dir}{The path the the git_stats.csv file}

\item{exclude_authors}{The list of authors to exclude from the stats}

\item{type}{The type of files to consider ("all", "R", "Rmd", or "Qmd")}

\item{tz}{The time zone to use for times}
}
\value{
A data frame with git stats data to be used by the learnitgrid Shiny
app.
}
\description{
Use data from a git_stats.csv file to get a history of commits
done in a repository for the learnitgrid Shiny app.
}
