% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_hook.R
\name{hook_last_chunk}
\alias{hook_last_chunk}
\title{Configure Knitr to generate a .Last.chunk object}
\usage{
hook_last_chunk()
}
\value{
The \code{opts_chunk} or the \code{knit_hook} set accordingly.
}
\description{
Change the Knitr renderer so that the last computed object is automatically
saved as \code{.Last.chunk}, and to record results from a chunk with chunk options
\verb{record=}.
}
