% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_grid_example.R
\name{install_grid_example}
\alias{install_grid_example}
\title{Install an extended example dataset to try the learnitgrid Shiny application}
\usage{
install_grid_example(dir = tempdir(), browse = FALSE)
}
\arguments{
\item{dir}{The directory where to decompress the extended example}

\item{browse}{Should we browse the example directory after decompression?}
}
\value{
The path to the example learnitgrid data is returned invisibly.
}
\description{
Install an extended example dataset to try the learnitgrid Shiny application
}
\examples{
# Install the extended examples in a temporary directory
\dontrun{
library(learnitgrid)
ex_dir <- install_grid_example(browse = TRUE)
}
}
\seealso{
\code{\link[=run_grid]{run_grid()}}
}
