% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitgrid-package.R
\docType{package}
\name{learnitgrid-package}
\alias{learnitgrid}
\alias{learnitgrid-package}
\title{learnitgrid: Shiny application for 'learnitdown' evaluation grids}
\description{
Shiny application to ease the correction and evaluation of GitHub projects with R/Rmd/Qmd files using evaluation grids in the context of 'learnitdown' exercises.
}
\details{
The {learnitgrid} package ... TODO: describe here
}
\section{Important functions}{

\itemize{
\item \code{\link[=create_context]{create_context()}} creates a context object (a list) with ...
\item \code{\link[=populate_table]{populate_table()}} computes the content for a \code{DT::datatable()} with the
evaluation grid information.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/learnitgrid}
  \item \url{https://www.sciviews.org/learnitgrid/}
  \item Report bugs at \url{https://github.com/SciViews/learnitgrid/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be}
}

}
\keyword{internal}
