% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_to_www}
\alias{link_to_www}
\alias{www_relative}
\title{Create symbolic links under www subdirectory or strip www subdirectory from the path}
\usage{
link_to_www(path, link)

www_relative(path)
}
\arguments{
\item{path}{The (absolute) path containing the documents.}

\item{link}{The name to use for the symbolic link under the www subdirectory of the Shiny app.}
}
\value{
\code{TRUE} if the symbolic link exists for \code{\link[=link_to_www]{link_to_www()}} or \code{FALSE}
otherwise. The modified path is returned by \code{\link[=www_relative]{www_relative()}}
}
\description{
Symbolic links in the www subdirectory of the Shiny app are
required for the application to properly display html documents or images in
html tags.
}
\section{Functions}{
\itemize{
\item \code{www_relative()}: Strip www/ in front of the relative paths for a Shiny app.

}}
