% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ref.R
\name{make_ref}
\alias{make_ref}
\alias{read_ref}
\title{Create or read reference files for tests}
\usage{
make_ref(
  name,
  ...,
  dir1 = fs::path(rprojroot::find_package_root_file(), "tests", "results"),
  dir2 = fs::path(rprojroot::find_package_root_file(), "tests", "reference"),
  nthreads = parallel::detectCores(logical = FALSE)
)

read_ref(
  name,
  ...,
  dir = fs::path(rprojroot::find_package_root_file(), "tests", "reference"),
  nthreads = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{name}{Name of the result file to transform into a reference file}

\item{...}{Further parameters passed to \code{\link[qs2:qd_read]{qs2::qd_read()}}}

\item{dir1}{Directory containing the result file}

\item{dir2}{Directory where to place the reference file}

\item{nthreads}{Number of threads to use for reading or writing the files}

\item{dir}{Directory containing the reference file}
}
\value{
The decoded content of the result file for \code{read_ref()} or the
number of bytes written for \code{make_ref()}.
}
\description{
Reference files allow to check if results are correct. They are not just a
copy of the result files. They are reencoded to avoid someone could just copy and paste from the reference to the results directories and cheat.
}
