% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_test}
\alias{make_test}
\alias{make_clean}
\alias{make_original}
\alias{make_solution}
\alias{make_prepare}
\title{Run action of a MAKEFILE from R}
\usage{
make_test()

make_clean()

make_original()

make_solution()

make_prepare()
}
\value{
The result from \code{make} is returned
}
\description{
The five possible actions are "test", to launch the test on the repository,
"clean" to delete several generated files, "original" to configure the
repository with original files, "solution" to configure the repository with
solution files, and "prepare" to prepare the repository for the final version
ready for the assignment in, say Github classroom.
}
