% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{prepare_files}
\alias{prepare_files}
\title{Prepare files for original or solution version}
\usage{
prepare_files(type = "original", remove_last_saved = FALSE, error = TRUE)
}
\arguments{
\item{type}{Either \code{"original"} or \code{"solution"}}

\item{remove_last_saved}{Should we also remove the last_saved version
(\code{FALSE} by default)?}

\item{error}{If \code{TRUE}, an error is thrown in case there is no file to
prepare}
}
\value{
\code{NULL}
}
\description{
Make sure all files are original or solution versions, and possibly also
remove last_saved versions. This is typically used to prepare the repository
for an assignment, or to switch from originals to solution to verify the
tests.
}
