% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_reporter.R
\name{project_reporter}
\alias{project_reporter}
\title{Create a testthat reporter suitable to test projects}
\usage{
project_reporter()
}
\value{
A testthat reporter
}
\description{
Create a testthat reporter suitable to test projects
}
