% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_object.R
\name{record_res}
\alias{record_res}
\alias{read_res}
\alias{write_res}
\alias{RO}
\alias{RN}
\alias{RODFS}
\alias{RNDFS}
\alias{ROMD5}
\alias{RNMD5}
\alias{ROP}
\alias{RNP}
\alias{ROA}
\alias{RNA}
\alias{ROSTR}
\alias{RNSTR}
\alias{ROCS}
\alias{RNCS}
\title{Record, read and write results that capture given characteristics of an object}
\usage{
record_res(
  object = ".Last.chunk",
  name = object,
  fun = NULL,
  ...,
  dir = fs::path(rprojroot::find_package_root_file(), "tests", "results"),
  env = parent.frame()
)

read_res(
  name,
  ...,
  dir = fs::path(rprojroot::find_package_root_file(), "tests", "results"),
  nthreads = parallel::detectCores(logical = FALSE)
)

write_res(
  object,
  name,
  ...,
  dir = fs::path(rprojroot::find_package_root_file(), "tests", "results"),
  nthreads = parallel::detectCores(logical = FALSE)
)

RO(
  object = ".Last.chunk",
  name = object,
  fun = NULL,
  ...,
  dir = fs::path(rprojroot::find_package_root_file(), "tests", "results"),
  env = parent.frame()
)

RN(name, object = ".Last.chunk", fun = NULL, ..., env = parent.frame())

RODFS(
  object = ".Last.chunk",
  name = object,
  fun = df_structure,
  ...,
  env = parent.frame()
)

RNDFS(
  name,
  object = ".Last.chunk",
  fun = df_structure,
  ...,
  env = parent.frame()
)

ROMD5(
  object = ".Last.chunk",
  name = object,
  fun = digest,
  ...,
  env = parent.frame()
)

RNMD5(name, object = ".Last.chunk", fun = digest, ..., env = parent.frame())

ROP(
  object = ".Last.chunk",
  part = "x",
  name = object,
  fun = object_part,
  ...,
  env = parent.frame()
)

RNP(
  name,
  part = "x",
  object = ".Last.chunk",
  fun = object_part,
  ...,
  env = parent.frame()
)

ROA(
  object = ".Last.chunk",
  attrib = "class",
  name = object,
  fun = object_attr,
  ...,
  env = parent.frame()
)

RNA(
  name,
  attrib = "class",
  object = ".Last.chunk",
  fun = object_attr,
  ...,
  env = parent.frame()
)

ROSTR(
  object = ".Last.chunk",
  part = "x",
  name = object,
  fun = object_str,
  ...,
  env = parent.frame()
)

RNSTR(
  name,
  part = "x",
  object = ".Last.chunk",
  fun = object_str,
  ...,
  env = parent.frame()
)

ROCS(
  object = ".Last.chunk",
  arg = "",
  name = object,
  fun = chart_structure,
  ...,
  env = parent.frame()
)

RNCS(
  name,
  arg = "",
  object = ".Last.chunk",
  fun = chart_structure,
  ...,
  env = parent.frame()
)
}
\arguments{
\item{object}{The object to record results from (\code{".Last.chunk"} by default,
which corresponds to the last R object produced in the current chunk in an
R Markdown or Quarto document)}

\item{name}{The name of the result file}

\item{fun}{The function used to compute results for the object}

\item{...}{Additional arguments to pass to \verb{fun=} or to the read/write
functions using \code{\link[qs2:qs_save]{qs2::qs_save()}} and \code{\link[qs2:qs_read]{qs2::qs_read()}}}

\item{dir}{The directory where to save the results}

\item{env}{The environment where to look for the object}

\item{nthreads}{The number of threads to use for reading and writing}

\item{part}{The part(s) (list item(s)) to use}

\item{attrib}{The attribute(s) to use}

\item{arg}{Additional argument to the function used to compute results.
Currently, could be \code{""} (by default) to record everything, or \code{"no.data"}
to ignore data in ggplot objects in \code{ROCS()} and \code{RNCS()}.}
}
\value{
The deserialized results for \code{read_res()} or the number of bytes
written invisibly for \code{write_res()}. \code{record_res()} invisibly returns
\code{TRUE} or \code{FALSE} depending on the success of the operation.
}
\description{
Record, read and write results that capture given characteristics of an object
}
\details{
The main function to record the results is \code{record_res()}. However, there are
several shortcuts to record specific characteristics of an object: \code{RO} is
the same as \code{record_res()}, \code{RN()} records by default \code{".Last.chunk"} but
focuses on the name of the result file, \code{RODFS()} records the main structure
of a data frame, \code{RNDFS()} is the same as \code{RODFS()} but focuses on the name
of the result file, \code{ROMD5()} records the MD5 hash of an object, \code{RNMD5()} is
the same focusing on the name, \code{ROP()} records one or several parts of an
object (items from a list), \code{RNP()} is the same focusing on the name, \code{ROA()}
records one or several attributes of the objects with \code{RNA()} focusing on the
name, \code{ROSTR()} records the \code{\link[utils:str]{utils::str()}} summary of the object, and
\code{RNSTR()} is the same focusing on the name of the result file.
}
