% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grid.R
\name{run_grid}
\alias{run_grid}
\title{Run the learnitgrid Shiny application}
\usage{
run_grid(data_dir = getOption("learnitgrid.data.dir", NULL))
}
\arguments{
\item{data_dir}{The directory containing the 'learnitgrid' data. If missing,
the "learnitgrid.data.dir" option is used. If \code{""}, the small example
dataset in this package is used.}
}
\value{
Nothing is returned, the function is used for its side-effect of
running the learnitgrid Shiny application.
}
\description{
Run the learnitgrid Shiny application
}
\examples{
\dontrun{
library(learnitgrid)
# Run a basic example in the learnitgrid package
run_grid("")
# Please, note that the .Rmd reports that are analyzed in this example
# are NOT included to save space, hence, error messages in corresponding
# columns of the evaluation grids.

# Install an extended example in tempdir() (inspect its content)
install_grid_example(browse = TRUE)
# ... and run the app with it
run_grid()
# The .Rmd reports are included in this example.
}
}
\seealso{
\code{\link[=install_grid_example]{install_grid_example()}}
}
