% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_answer.R
\name{select_answer}
\alias{select_answer}
\alias{obfuscate}
\alias{get_word}
\title{A simple multiple choice system in a R chunk, compatible with git and R Markdown or Quarto documents}
\usage{
select_answer(x, name = NULL)

obfuscate(x)

get_word(x)
}
\arguments{
\item{x}{A String with the different answers, starting with []. The user has
to "check" the correct items by adding 'x' or 'X' inside the brackets}

\item{name}{The name of the select_answer. If not provided, the label of the
chunk where the \code{select_answer()} function is placed.}
}
\value{
A Markdown paragraph containing only the selected items (and the
selected items are also written in a result file)
}
\description{
A simple multiple choice system in a R chunk, compatible with git and R
Markdown or Quarto documents. \code{obfuscate()} and \code{get_word()} are used to hide
correct answers in the {testthat} tests.
}
