% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_to.R
\name{switch_to_original}
\alias{switch_to_original}
\alias{switch_to_solution}
\alias{switch_to_last_saved}
\alias{save_as_original}
\alias{save_as_solution}
\title{Switch to a given version of the document or save such a version}
\usage{
switch_to_original(file = NULL, error = TRUE)

switch_to_solution(file = NULL, error = TRUE)

switch_to_last_saved(file = NULL, error = TRUE)

save_as_original(file = NULL)

save_as_solution(file = NULL)
}
\arguments{
\item{file}{The file to switch to a version. If not provided, the currently
edited file in RStudio is used by default}

\item{error}{In case the version does not exists, do we issue an error or not
(no by default)}
}
\value{
\code{NULL} is returned invisibly. The functions are used for their
side-effect of switching document versions.
}
\description{
When there are several versions of a document (R script, R
Markdown or Quarto document), switch the main file to one of the saved
versions: original version, solution with the correction, or last saved
version. The \code{save_as_original()} and \code{save_as_solution()} functions are
used to create respectively the original and solution versions of the
document. These files have same name as the initial file but ending with
\verb{_original}, \verb{_solution} or \verb{_last_saved}., and starting with a dot \code{.}.
They are thus "hidden" files. For R scripts, the extension is also changed
into \code{.Rscript} for the original and solution versions.
}
