% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dir.R
\name{test_dir}
\alias{test_dir}
\title{Test a project directory, possibly limit the number of uses}
\usage{
test_dir(path, reporter = project_reporter(), times = 10L, ...)
}
\arguments{
\item{path}{The path to test}

\item{reporter}{The testthat reporter to use}

\item{times}{The maximum number of times the tests can be run by the end-user}

\item{...}{Additional arguments to pass to \code{\link[testthat:test_dir]{testthat::test_dir()}}}
}
\value{
The result of \code{\link[testthat:test_dir]{testthat::test_dir()}}
}
\description{
Test a project directory, possibly limit the number of uses
}
