% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitgrid-package.R
\docType{package}
\name{learnitgrid-package}
\alias{learnitgrid}
\alias{learnitgrid-package}
\title{learnitgrid: Manage Rubrics or Assessment Grids for GitHub Repositories}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create and manage semi-automatically rubrics to assess GitHub projects (R scripts, R Markdown or Quarto files). Create directed projects where students have to complete documents and submit them to GitHub (classroom) so that they are evaluated using the rubric (or assessment grid).
}
\details{
The learnitgrid package allows to create and manage semi-automatically
rubrics to assess GitHub projects (R scripts, R Markdown or Quarto files).
Create directed projects where students have to complete documents and submit
them to GitHub (classroom) so that they are evaluated using the rubric
(or assessment grid).
}
\section{Important functions}{

\itemize{
\item \code{\link[=install_grid_example]{install_grid_example()}} decompress an example dataset.
\item \code{\link[=run_grid]{run_grid()}} run the learnitgrid Shiny application.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/learnitr/learnitgrid}
  \item \url{https://learnitr.github.io/learnitgrid/}
  \item \url{https://learnitr.r-universe.dev/learnitgrid}
  \item Report bugs at \url{https://github.com/learnitr/learnitgrid/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID})
}

}
\keyword{internal}
